delete comment

+++ backend/controllers/postController.js
@@ -220,12 +220,17 @@ const repost = asyncHandler(async (req, res) => {
 const deletePost = asyncHandler(async (req, res) => {
   const post = await Post.findById(req.params.id)
 
-  // console.log(post.user._id == req.user.id)
-  // console.log(post.user)
-  // console.log(req.user._id)
-
-  await Post.deleteOne({ _id: post._id })
-  res.status(200)
+  if (post) {
+    if (post.parent) {
+      await Post.findByIdAndUpdate(post.parent, {
+        $pull: {
+          comments: post._id,
+        },
+      })
+    }
+    await Post.deleteOne({ _id: post._id })
+    res.status(200)
+  }
 })
 
 export {
